feat: public search

+++ main-service/src/main/java/ru/practicum/event/dto/EventFullDto.java
@@ -30,5 +30,5 @@ public class EventFullDto {
     Boolean requestModeration;
     EventState state;
     String title;
-    Integer views;
+    Long views;
 }
