+++ main-service/src/main/java/ru/practicum/event/dto/EventShortDto.java
@@ -20,5 +20,5 @@ public class EventShortDto {
     UserDto initiator;
     Boolean paid;
     String title;
-    Integer views;
+    Long views;
 }
